// src/extractor/mdbook-line-type.ts
export interface INumberedLine {
  title: string;
  path: string | null;
  spacesNumber: number;
}

export interface IPartLine {
  title: string;
  path: string | null;
}

export interface IXfixLine {
  title: string;
  path: string | null;
}

export interface IPartLine {
  title: string;
  path: string | null;
}

export interface ITitleLine {
  title: string;
}